Remove ksort code and clean up chromap header.
